K8s cluster deployment and testing

+++ vmbc-ethereum-starter-kit/release-1.6/starter-kit/vmbc/script/k8s-destroy.sh
@@ -19,9 +19,11 @@ cd $CWD/../config
 rm -rf devdata
 cd - 
 
-echo ''
-echo '---------------- Delete .env ----------------'
-rm ../.env.config
+if [ -f ../.env.config ]; then
+   echo ''
+   echo '---------------- Delete .env ----------------'
+   rm ../.env.config
+fi
 
 echo ''
 echo '---------------- Delete tmp files ----------------'
